better PR diff
